refactor(history): Expose mergeDuplicatePeers from DbUpgrader
To be used by both dbSchema8to9 and dbTo11.
